+++ src/commonMain/kotlin/com/hoc081098/flowext/bufferCount.kt
@@ -1,7 +1,7 @@
 package com.hoc081098.flowext
 
+import kotlinx.coroutines.InternalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.flow
 
 /**
@@ -13,6 +13,7 @@ import kotlinx.coroutines.flow.flow
  * For example if [startBufferEvery] is 2, then a new buffer will be started on every other value from the source.
  * A new buffer is started at the beginning of the source by default.
  */
+@InternalCoroutinesApi // TODO: Remove InternalCoroutinesApi (https://github.com/Kotlin/kotlinx.coroutines/issues/3078)
 public fun <T> Flow<T>.bufferCount(
   bufferSize: Int,
   startBufferEvery: Int? = null,
@@ -26,6 +27,7 @@ public fun <T> Flow<T>.bufferCount(
   }
 }
 
+@InternalCoroutinesApi // TODO: Remove InternalCoroutinesApi (https://github.com/Kotlin/kotlinx.coroutines/issues/3078)
 private fun <T> Flow<T>.bufferSkip(bufferSize: Int, skip: Int): Flow<List<T>> {
   return flow {
     val buffers = ArrayDeque<MutableList<T>>()
@@ -70,6 +72,7 @@ private fun <T> Flow<T>.bufferSkip(bufferSize: Int, skip: Int): Flow<List<T>> {
   }
 }
 
+@InternalCoroutinesApi // TODO: Remove InternalCoroutinesApi (https://github.com/Kotlin/kotlinx.coroutines/issues/3078)
 private fun <T> Flow<T>.bufferExact(bufferSize: Int): Flow<List<T>> {
   return flow {
     var buffer: MutableList<T> = mutableListOf()