+++ src/commonMain/kotlin/com/hoc081098/flowext/timer.kt
@@ -4,7 +4,6 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlin.time.Duration
-import kotlin.time.ExperimentalTime
 
 /**
  * Creates a [Flow] that will wait for a specified time, before emitting the [value].
@@ -17,7 +16,6 @@ public fun <T> timer(value: T, timeMillis: Long): Flow<T> = flow {
 /**
  * Creates a [Flow] that will wait for a given [duration], before emitting the [value].
  */
-@ExperimentalTime
 public fun <T> timer(value: T, duration: Duration): Flow<T> = flow {
   delay(duration)
   emit(value)