+++ src/commonMain/kotlin/com/hoc081098/flowext/withLatestFrom.kt
@@ -1,6 +1,7 @@
 package com.hoc081098.flowext
 
 import kotlinx.coroutines.CoroutineStart
+import kotlinx.coroutines.InternalCoroutinesApi
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.onSuccess
 import kotlinx.coroutines.coroutineScope
@@ -16,6 +17,7 @@ import kotlinx.coroutines.launch
  * @param other Second [Flow]
  * @param transform A transform function to apply to each value from self combined with the latest value from the second [Flow], if any.
  */
+@InternalCoroutinesApi // TODO: Remove InternalCoroutinesApi (https://github.com/Kotlin/kotlinx.coroutines/issues/3078)
 public fun <A, B, R> Flow<A>.withLatestFrom(
   other: Flow<B>,
   transform: suspend (A, B) -> R
@@ -25,7 +27,7 @@ public fun <A, B, R> Flow<A>.withLatestFrom(
       val otherValues = Channel<Any>(Channel.CONFLATED)
       launch(start = CoroutineStart.UNDISPATCHED) {
         other.collect {
-          return@collect otherValues.send(it ?: NullValue)
+          return@collect otherValues.send(it ?: NULL_VALUE)
         }
       }
 
@@ -38,7 +40,7 @@ public fun <A, B, R> Flow<A>.withLatestFrom(
         emit(
           transform(
             value,
-            NullValue.unbox(lastValue ?: return@collect)
+            NULL_VALUE.unbox(lastValue ?: return@collect)
           ),
         )
       }
@@ -46,6 +48,7 @@ public fun <A, B, R> Flow<A>.withLatestFrom(
   }
 }
 
+@InternalCoroutinesApi // TODO: Remove InternalCoroutinesApi (https://github.com/Kotlin/kotlinx.coroutines/issues/3078)
 @Suppress("NOTHING_TO_INLINE")
 public inline fun <A, B> Flow<A>.withLatestFrom(other: Flow<B>): Flow<Pair<A, B>> =
   withLatestFrom(other) { a, b -> a to b }