+++ src/commonTest/kotlin/com/hoc081098/flowext/RetryWhenWithDelayStrategyTest.kt
@@ -11,9 +11,7 @@ import kotlin.test.assertEquals
 import kotlin.test.assertIs
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
-import kotlin.time.ExperimentalTime
 
-@ExperimentalTime
 @ExperimentalCoroutinesApi
 @InternalCoroutinesApi
 class RetryWhenWithDelayStrategyTest {